+++ src/ome_zarr_converters_tools/fractal/_compute_task.py
@@ -116,9 +116,8 @@ def generic_compute_task(
             f"{init_args.tiled_image_json_dump_url}"
         )
     registration_pipeline = build_default_registration_pipeline(
-        alignment_corrections=init_args.converter_options.alignment_correction,
-        tiling_mode=init_args.converter_options.tiling_mode,
-        tolerance=init_args.converter_options.tiling_tolerance,
+        alignment_corrections=init_args.converter_options.stage_position_corrections,
+        tiling_strategy=init_args.converter_options.tiling_strategy,
     )
     ome_zarr = tiled_image_creation_pipeline(
         zarr_url=zarr_url,